AI-Jobs-Portal/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
openai==0.28.0
email-validator==2.0.0
Pillow==10.0.0